src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
cachetools==5.3.0
    # via google-auth
celery==5.2.7
    # via
    #   -r src/requirements/requirements.in
    #   django-celery-beat
    #   django-celery-results
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==39.0.0
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.0.9
    # via
    #   -r src/requirements/requirements.in
    #   django-allauth
    #   django-celery-beat
    #   django-htmx
django-allauth==0.52.0
    # via -r src/requirements/requirements.in
django-celery-beat==2.4.0
    # via -r src/requirements/requirements.in
django-celery-results==2.4.0
    # via -r src/requirements/requirements.in
django-htmx==1.13.0
    # via -r src/requirements/requirements.in
django-timezone-field==5.0
    # via django-celery-beat
faker==16.6.1
    # via -r src/requirements/requirements.in
google-api-core[grpc]==2.11.0
    # via google-cloud-secret-manager
google-auth==2.16.0
    # via google-api-core
google-cloud-secret-manager==2.15.1
    # via -r src/requirements/requirements.in
googleapis-common-protos[grpc]==1.58.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.6
    # via google-cloud-secret-manager
grpcio==1.51.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.51.1
    # via google-api-core
idna==3.4
    # via requests
joblib==1.2.0
    # via scikit-surprise
kombu==5.2.4
    # via celery
numpy==1.24.1
    # via
    #   -r src/requirements/requirements.in
    #   scikit-surprise
    #   scipy
oauthlib==3.2.2
    # via requests-oauthlib
prompt-toolkit==3.0.36
    # via click-repl
proto-plus==1.22.2
    # via google-cloud-secret-manager
protobuf==4.21.12
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via django-allauth
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   faker
    #   python-crontab
python-decouple==3.7
    # via -r src/requirements/requirements.in
python3-openid==3.2.0
    # via django-allauth
pytz==2022.7.1
    # via
    #   celery
    #   django-timezone-field
redis==4.4.2
    # via -r src/requirements/requirements.in
requests==2.28.2
    # via
    #   django-allauth
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
rsa==4.9
    # via google-auth
scikit-surprise==1.1.3
    # via -r src/requirements/requirements.in
scipy==1.10.0
    # via scikit-surprise
six==1.16.0
    # via
    #   click-repl
    #   google-auth
    #   python-dateutil
sqlparse==0.4.3
    # via django
tzdata==2022.7
    # via django-celery-beat
urllib3==1.26.14
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
